// src/school/interfaces/school.interface.ts
export interface SchoolInterface {
  schoolId?: string
  schoolName?: string
  email?: string
  schoolRefId?: string
  instituteManagement?: string
  address?: string
  schoolType?: string
  website?: string
  street?: string
  city?:string
  district?: string
  state?: string
  pincode?: string
  }